Introducing extremely simplified dispatching and consuming messages. Getting rid of not needed any more code.
