+++ composer.json
@@ -13,9 +13,6 @@
         "doctrine/doctrine-bundle": "^2.3",
         "doctrine/doctrine-migrations-bundle": "^3.0",
         "doctrine/orm": "^2.8",
-        "gedmo/doctrine-extensions": "^3.0",
-        "guzzlehttp/guzzle": "^7.3",
-        "illuminate/http": "^8.34",
         "nesbot/carbon": "^2.46",
         "symfony/console": "5.*.*",
         "symfony/doctrine-messenger": "5.*.*",
