+++ config/packages/doctrine.yaml
@@ -4,6 +4,8 @@ doctrine:
         types:
             datetime_immutable: \Carbon\Doctrine\DateTimeImmutableType
             datetime: \Carbon\Doctrine\DateTimeType
+            carbon_immutable: \Carbon\Doctrine\DateTimeImmutableType
+            carbon: \Carbon\Doctrine\DateTimeType
 
         # IMPORTANT: You MUST configure your server version,
         # either here or in the DATABASE_URL env var (see .env file)
@@ -15,7 +17,7 @@ doctrine:
         mappings:
             App:
                 is_bundle: false
-                type: annotation
+                type: attribute
                 dir: '%kernel.project_dir%/src/Entity'
                 prefix: 'App\Entity'
                 alias: App
