+++ config/services.yaml
@@ -29,12 +29,3 @@ services:
 
     # add more service definitions when explicit configuration is needed
     # please note that last definitions always *replace* previous ones
-
-    App\Services\TrackingResolver:
-        public: true
-        arguments:
-            $apiUri: '%env(resolve:TRACKING_API_URI)%'
-
-
-    # illuminate http
-    Illuminate\Http\Client\Factory:
