+++ docker/php/etc/supervisor/conf.d/emulatecron.conf
@@ -1,6 +1,6 @@
-[program:app_trackings_manage]
-command= /bin/sh -c "while [ true ]; do (/app/bin/console app:trackings:manage -vv); sleep 10; done"
-numprocs=1
-stdout_logfile=/dev/fd/1
-stdout_logfile_maxbytes=0
-autorestart=true
+;[program:app_trackings_manage]
+;command= /bin/sh -c "while [ true ]; do (/app/bin/console app:trackings:manage -vv); sleep 10; done"
+;numprocs=1
+;stdout_logfile=/dev/fd/1
+;stdout_logfile_maxbytes=0
+;autorestart=true
