+++ docker/php/etc/supervisor/conf.d/log.conf
@@ -1,4 +1,4 @@
-[program:symfony-dev.log]
+[program:log]
 command=/usr/bin/tail -f /app/var/log/dev.log
 stdout_logfile=/dev/fd/1
 stdout_logfile_maxbytes=0
