+++ docker/php/etc/supervisor/conf.d/messenger.conf
@@ -1,10 +1,10 @@
-[program:messenger-consume]
-command=php /app/bin/console messenger:consume async --time-limit=30 --limit=5
-numprocs=4
-startsecs=0
-autostart=true
-autorestart=true
-process_name=%(program_name)s_%(process_num)02d
+;[program:messenger-consume]
+;command=php /app/bin/console messenger:consume async --time-limit=30 --limit=5
+;numprocs=4
+;startsecs=0
+;autostart=true
+;autorestart=true
+;process_name=%(program_name)s_%(process_num)02d
 ; stdout_logfile=/dev/stdout
 ; stdout_logfile_maxbytes=0
-autorestart=true
+;autorestart=true
