+++ docker/php/etc/supervisor/supervisord.conf
@@ -26,8 +26,5 @@ serverurl=unix:///var/run/supervisor.sock ; use a unix:// URL  for a unix socket
 ; interpreted as relative to this file.  Included files *cannot*
 ; include files themselves.
 
-; [program:php]
-; command=docker-php-entrypoint php-fpm # This is our main process so let us put it here
-
 [include]
 files = /etc/supervisor/conf.d/*.conf
