--- src/Entity/VictoriousGnome.php
@@ -1,41 +1,0 @@
-<?php
-
-namespace App\Entity;
-
-use App\Repository\VictoriousGnomeRepository;
-use Doctrine\ORM\Mapping as ORM;
-
-/**
- * @ORM\Entity(repositoryClass=VictoriousGnomeRepository::class)
- */
-class VictoriousGnome
-{
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
-     */
-    private $id;
-
-    /**
-     * @ORM\Column(type="string", length=255)
-     */
-    private $foo;
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    public function getFoo(): ?string
-    {
-        return $this->foo;
-    }
-
-    public function setFoo(string $foo): self
-    {
-        $this->foo = $foo;
-
-        return $this;
-    }
-}
