+++ symfony.lock
@@ -1,7 +1,4 @@
 {
-    "behat/transliterator": {
-        "version": "v1.3.0"
-    },
     "composer/package-versions-deprecated": {
         "version": "1.11.99.1"
     },
@@ -87,39 +84,6 @@
     "friendsofphp/proxy-manager-lts": {
         "version": "v1.0.3"
     },
-    "gedmo/doctrine-extensions": {
-        "version": "v3.0.3"
-    },
-    "guzzlehttp/guzzle": {
-        "version": "7.3.0"
-    },
-    "guzzlehttp/promises": {
-        "version": "1.4.1"
-    },
-    "guzzlehttp/psr7": {
-        "version": "1.8.1"
-    },
-    "illuminate/collections": {
-        "version": "v8.34.0"
-    },
-    "illuminate/contracts": {
-        "version": "v8.34.0"
-    },
-    "illuminate/filesystem": {
-        "version": "v8.34.0"
-    },
-    "illuminate/http": {
-        "version": "v8.34.0"
-    },
-    "illuminate/macroable": {
-        "version": "v8.34.0"
-    },
-    "illuminate/session": {
-        "version": "v8.34.0"
-    },
-    "illuminate/support": {
-        "version": "v8.34.0"
-    },
     "laminas/laminas-code": {
         "version": "4.0.0"
     },
@@ -141,23 +105,8 @@
     "psr/event-dispatcher": {
         "version": "1.0.0"
     },
-    "psr/http-client": {
-        "version": "1.0.1"
-    },
-    "psr/http-factory": {
-        "version": "1.0.1"
-    },
-    "psr/http-message": {
-        "version": "1.0.1"
-    },
     "psr/log": {
         "version": "1.1.3"
-    },
-    "psr/simple-cache": {
-        "version": "1.0.1"
-    },
-    "ralouphie/getallheaders": {
-        "version": "3.0.3"
     },
     "symfony/amqp-messenger": {
         "version": "v5.2.4"
@@ -272,9 +221,6 @@
             "config/packages/messenger.yaml"
         ]
     },
-    "symfony/mime": {
-        "version": "v5.2.5"
-    },
     "symfony/monolog-bridge": {
         "version": "v5.2.5"
     },
@@ -294,9 +240,6 @@
         ]
     },
     "symfony/polyfill-intl-grapheme": {
-        "version": "v1.22.1"
-    },
-    "symfony/polyfill-intl-idn": {
         "version": "v1.22.1"
     },
     "symfony/polyfill-intl-normalizer": {
@@ -373,8 +316,5 @@
     },
     "symfony/yaml": {
         "version": "v5.2.5"
-    },
-    "voku/portable-ascii": {
-        "version": "1.5.6"
     }
 }
